aws/requirements.txt
requests==2.19.1
grpcio
oauth2client==4.1.2
protobuf==3.6.0
Pillow
python-Levenshtein
fuzzywuzzy
boto3
awscli
pandas
numpy